Move request listening and attribute key extraction to app scope
