+++ src/components/Debugger/Timeline/index.tsx
@@ -1,11 +1,15 @@
-import type { Cookie, Entry } from "har-format";
 import { h, type FunctionComponent } from "preact";
-import { useEffect, useMemo, useRef, useState } from "preact/hooks";
+import { useMemo, useRef, useState } from "preact/hooks";
 
 import { endpointAnalytics, trackerAnalytics } from "../../../ts/analytics";
-import { esMap, protocol } from "../../../ts/protocol";
-import type { IBeaconSummary, ITimeline } from "../../../ts/types";
-import { b64d, colorOf, hash, tryb64 } from "../../../ts/util";
+import { GA_REQUIRED_FIELDS } from "../../../ts/extractBatchContents";
+import { protocol } from "../../../ts/protocol";
+import type {
+  BatchContents,
+  IBeaconSummary,
+  ITimeline,
+} from "../../../ts/types";
+import { b64d, colorOf, tryb64 } from "../../../ts/util";
 
 import { EventEntry } from "./EventEntry";
 import { PageGroup } from "./PageGroup";
@@ -13,7 +17,6 @@ import { TimelineChrome } from "./TimelineChrome";
 
 import "./Timeline.css";
 
-const GA_REQUIRED_FIELDS = ["tid", "cid", "t", "v", "_gid"] as const;
 const KNOWN_FAKE_PAGES = [
   "badbucket.invalid",
   "elasticsearch.invalid",
@@ -149,182 +152,25 @@ const summariseBatch = ({
   return results;
 };
 
-const getPageUrl = (entries: Entry[]) => {
-  const page = entries
-    .flatMap((entry) => entry.request.headers)
-    .find((header) => /referr?er/i.test(header.name));
-  return page && page.value ? new URL(page.value) : null;
-};
-
-const extractNetworkUserId = (cookies: Cookie[]): Cookie | undefined => {
-  // consider only cookies with the UUID format
-  const uuidRegexp =
-    /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
-  const uuidCookies = cookies.filter(({ value }) => uuidRegexp.test(value));
-  // prefer a cookie with the name `sp` or take the first one
-  return uuidCookies.find((x) => x.name === "sp") ?? uuidCookies.shift();
-};
-
-type BatchContents = {
-  id: string;
-  collector: string;
-  collectorPath: string;
-  method: string;
-  pageref?: string;
-  events: Map<string, string>[];
-  serverAnonymous: boolean;
-  status: number;
-  statusText: string;
-};
-
-const extractBatchContents = (
-  { request, response, pageref, startedDateTime }: Entry,
-  index: number,
-): BatchContents => {
-  const { headers, method, postData, url } = request;
-  const ref = pageref && /^page_\d+/.test(pageref) ? "beacon" : pageref;
-  const id = ref + hash(new Date(startedDateTime).toJSON() + url + index);
-
-  const collectorUrl = new URL(url);
-  const collector = collectorUrl.hostname;
-  const collectorPath = collectorUrl.pathname;
-  const events: Map<string, string>[] = [];
-
-  // These fields are set by enrich based on the headers sent to/from the collector
-  // Extract these manually so we can display them if they aren't explicitly set
-  const nuid = extractNetworkUserId(response.cookies)?.value;
-  const ua = headers.find(
-    ({ name }) => name.toLowerCase() === "user-agent",
-  )?.value;
-  const lang = headers.find(
-    ({ name }) => name.toLowerCase() === "accept-language",
-  )?.value;
-  const refr = headers.find(
-    ({ name }) => name.toLowerCase() === "referer",
-  )?.value;
-  const cl = headers.find(
-    ({ name }) => name.toLowerCase() === "content-length",
-  )?.value;
-  const ct = headers.find(
-    ({ name }) => name.toLowerCase() === "content-type",
-  )?.value;
-
-  const serverAnonymous = headers.some(
-    ({ name, value }) => name.toLowerCase() === "sp-anonymous" && value === "*",
-  );
-
-  if (method.toUpperCase() === "POST") {
-    if (postData == null && cl != null && parseInt(cl, 10) > 0) {
-      const event = new Map(
-        Object.entries({
-          error: [
-            "A post request to a known tracker URL was seen, but we could not extract the request body.",
-            "This is usually the result of a request sent with the `beacon` eventMethod",
-            "as a page is unloaded (e.g. a link click or form submission event that navigates to a new page).",
-            "The browser is unable to supply the request body without a reference to the page, so we can't display the events.",
-            "Although we can not inspect the event, it was probably correctly sent to the collector.",
-          ].join(" "),
-          e: "Unknown Event",
-          endpoint: url,
-          payloadSize: cl,
-          contentType: ct ?? "",
-          stm: String(+new Date()),
-        }),
-      );
-
-      events.push(event);
-    } else if (postData && postData.text) {
-      try {
-        const payload: unknown = JSON.parse(postData.text);
-
-        if (!payload || typeof payload !== "object")
-          throw Error("Expected Snowplow payload_data SDJ, got non-object");
-        if (!("schema" in payload && "data" in payload))
-          throw Error("Expected Snowplow payload_data SDJ");
-        if (
-          typeof payload.schema !== "string" ||
-          !/\/payload_data\//.test(payload.schema)
-        )
-          throw Error("Unexpectedly got non-payload_data SDJ");
-        if (!Array.isArray(payload.data))
-          throw Error("Invalid payload_data object without array for data");
-
-        for (const pl of payload.data) {
-          events.push(new Map(Object.entries(pl)));
-        }
-      } catch (jsonErr) {
-        try {
-          const ga = postData.text
-            .split("\n")
-            .map((line) => new Map(new URLSearchParams(line)));
-
-          const validGa = ga.filter((b) =>
-            GA_REQUIRED_FIELDS.every((k) => b.has(k)),
-          );
-          events.push(...validGa);
-        } catch (urlErr) {
-          console.error("Invalid request payload", request, [jsonErr, urlErr]);
-        }
-      }
-    } else {
-      console.error("Unexpected empty body in request", request);
-    }
-  } else {
-    // GET request, parse the params
-    const event = new Map(collectorUrl.searchParams);
-
-    if (lang && !event.has("lang")) {
-      const langval = /^[^;,]+/.exec(lang);
-      event.set("lang", langval ? langval[0] : lang);
-    }
-
-    events.push(event);
-  }
-
-  events.forEach((event) => {
-    if (nuid && !event.has("nuid")) {
-      event.set("nuid", nuid);
-    }
-    if (ua && !event.has("ua")) {
-      event.set("ua", ua);
-    }
-    if (lang && !event.has("lang")) {
-      event.set("lang", lang);
-    }
-    if (refr && !event.has("url")) {
-      event.set("url", refr);
-    }
-  });
-
-  return {
-    id,
-    collector,
-    collectorPath,
-    method,
-    pageref: ref,
-    events,
-    serverAnonymous,
-    status: response.status,
-    statusText: (response as any)._error || response.statusText,
-  };
+const getPageUrl = (batches: BatchContents[]) => {
+  const page = batches.find(({ sendingPage }) => sendingPage)?.sendingPage;
+  return page ? new URL(page) : null;
 };
 
 export const Timeline: FunctionComponent<ITimeline> = ({
   active,
   batches,
   destinationManager,
+  requestsRef,
   resolver,
   setActive,
   setApp,
-  setAttributeKeys,
-  setEventCount,
   setModal,
   addRequests,
   clearRequests,
 }) => {
   const [filterStr, setFilterStr] = useState<string>("");
 
-  const batchRef = useRef<Entry[]>([]);
   const summariesRef = useRef<IBeaconSummary[][]>([]);
 
   const filter = useMemo(() => {
@@ -337,53 +183,15 @@ export const Timeline: FunctionComponent<ITimeline> = ({
 
   const fallbackUrl = useMemo(() => getPageUrl(batches), [batches]);
 
-  const batchContents = useMemo(
-    () => batches.map(extractBatchContents),
-    [batches],
-  );
-
-  setEventCount(
-    batchContents.reduce(
-      (sum: number | undefined = 0, batch) => sum + batch.events.length,
-      undefined,
-    ),
-  );
-
   const batchSummaries = useMemo(
     () =>
-      batchContents.map((batch) =>
+      batches.map((batch) =>
         summariseBatch(batch).filter((summary) =>
           filterRequest(summary, filter),
         ),
       ),
-    [batchContents, filter],
+    [batches, filter],
   );
-
-  useEffect(() => {
-    setAttributeKeys((targets) => {
-      let dirty = false;
-
-      for (const [attributeKey, identifiers] of Object.entries(targets)) {
-        const payloadKey =
-          attributeKey in esMap
-            ? esMap[attributeKey as keyof typeof esMap]
-            : undefined;
-        if (!payloadKey) continue;
-
-        for (const batch of batchSummaries) {
-          for (const { payload } of batch) {
-            const id = payload.get(payloadKey);
-            if (id != null) {
-              dirty = dirty || !identifiers.has(id);
-              identifiers.add(id);
-            }
-          }
-        }
-      }
-
-      return dirty ? { ...targets } : targets;
-    });
-  }, [batchSummaries]);
 
   const pageGroups = useMemo(() => {
     const groups: [string, IBeaconSummary[]][] = [];
@@ -410,15 +218,14 @@ export const Timeline: FunctionComponent<ITimeline> = ({
     return groups;
   }, [batchSummaries, fallbackUrl]);
 
-  batchRef.current = batches;
   summariesRef.current = batchSummaries;
   if (pageGroups.length && !active) setActive(pageGroups[0][1][0]);
 
   return (
     <TimelineChrome
-      addRequests={addRequests}
       active={active}
-      batchRef={batchRef}
+      addRequests={addRequests}
+      requestsRef={requestsRef}
       summariesRef={summariesRef}
       clearRequests={clearRequests}
       destinationManager={destinationManager}
